refactor(core): Reorganizing internal/exported utils
