+++ sdk/core/package.json
@@ -19,10 +19,6 @@
     "./sdk/runtime": {
       "types": "./src/sdk/runtime/index.d.ts",
       "import": "./src/sdk/runtime/index.js"
-    },
-    "./sdk_utils": {
-      "types": "./src/utils/index.d.ts",
-      "import": "./src/utils/index.js"
     }
   },
   "files": [
@@ -73,8 +69,7 @@
     "#consts": "./src/consts.js",
     "#errors": "./src/errors.js",
     "#logger": "./src/logger/index.js",
-    "#utils": "./src/utils/index.js",
-    "#internal_utils/*": "./src/internal_utils/*.js",
+    "#helpers/*": "./src/helpers/*.js",
     "#tracing": "./src/tracing/internal_interface.js",
     "#trace_attribute": "./src/tracing/trace_attribute.js",
     "#async_storage": "./src/async_storage.js",
