+++ sdk/core/src/helpers/errors.js
