+++ sdk/core/src/helpers/errors.spec.js
