+++ sdk/core/src/helpers/trace_info.js
