+++ sdk/core/src/helpers/trace_info.spec.js
