+++ sdk/core/src/sdk/helpers/index.d.ts
@@ -9,3 +9,4 @@
  */
 export * from './component_metadata.js';
 export * from './path.js';
+export * from './objects.js';
