+++ sdk/core/src/sdk/helpers/index.js
@@ -1,2 +1,3 @@
 export * from './component_metadata.js';
 export * from './path.js';
+export * from './objects.js';
