+++ sdk/core/src/sdk/helpers/objects.js
@@ -1,0 +1,8 @@
+import { clone, deepMerge, deepMergeWithResolver, isPlainObject } from '#helpers/object';
+
+export const Objects = {
+  clone,
+  deepMerge,
+  deepMergeWithResolver,
+  isPlainObject
+};
