--- sdk/core/src/utils/index.js
@@ -1,1 +1,0 @@
-export * from './utils.js';
