+++ sdk/core/src/worker/configs.js
@@ -1,5 +1,5 @@
 import * as z from 'zod';
-import { isStringboolTrue } from '#utils';
+import { isStringboolTrue } from '#helpers/string';
 
 class InvalidEnvVarsErrors extends Error { }
