+++ sdk/core/src/worker/interceptors/workflow.js
@@ -1,10 +1,10 @@
 // THIS RUNS IN THE TEMPORAL'S SANDBOX ENVIRONMENT
 import { workflowInfo, proxySinks, ContinueAsNew, isCancellation } from '@temporalio/workflow';
 import { memoToHeaders } from './headers.js';
-import { deepMerge } from '#utils';
-import { buildApplicationFailureWithDetails } from '#internal_utils/errors';
+import { deepMerge } from '#helpers/object';
+import { buildApplicationFailureWithDetails } from '#helpers/errors';
 import { METADATA_ACCESS_SYMBOL, WorkflowSpecialOutput } from '#consts';
-import { createWorkflowDetails } from '#internal_utils/temporal_context';
+import { createWorkflowDetails } from '#helpers/temporal_context';
 
 // this is a dynamic generated file with activity configs overwrites
 import stepOptions from '../temp/__activity_options.js';
