+++ sdk/credentials/src/credentials.ts
@@ -1,7 +1,7 @@
 import { dirname } from 'node:path';
 import { MissingCredentialError } from './errors.js';
 import { Context } from '@outputai/core/sdk/runtime';
-import { deepMerge } from '@outputai/core/sdk_utils';
+import { Objects } from '@outputai/core/sdk/helpers';
 import { getProvider } from './provider_registry.js';
 
 const getNestedValue = ( obj: Record<string, unknown>, dotPath: string ): unknown =>
@@ -37,7 +37,7 @@ const loadForWorkflow = ( workflowName: string, workflowDir: string | undefined 
     workflowDir,
     environment: detectEnvironment()
   } );
-  const merged = workflowData ? deepMerge( globalData, workflowData ) as Record<string, unknown> : globalData;
+  const merged = workflowData ? Objects.deepMerge( globalData, workflowData ) as Record<string, unknown> : globalData;
   cache.set( workflowName, merged );
   return merged;
 };
